pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "casic"
version = "0.1.0"
description = "CASIC - CAN Stack Integrity Checker"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "CASIC Contributors" }]
license = "Apache-2.0"
classifiers = [
	"License :: OSI Approved :: Apache Software License",
]
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
can = ["python-can>=4.3"]
test = ["pytest>=8.0"]

[project.scripts]
casic = "casic.cli.config_runner:main"
cansic = "casic.cli.main:main_cansic"
udsic = "casic.cli.main:main_udsic"
j1939sic = "casic.cli.main:main_j1939sic"
cosic = "casic.cli.main:main_cosic"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
include = ["casic*"]
